2025_fall/outrun/requirements.txt
pygame
PyOpenGL
numpy
Pillow
